# store/proxy: Deduplicate chunks on StoreAPI level. Recommend chunk sorting for StoreAPI.