Blocked by: https://github.com/thanos-io/thanos/pull/2548

Partially fixes: https://github.com/thanos-io/thanos/issues/2303

It should make certain calls where we have overlapped data from StoreAPIs much faster and more efficient.

## Changes: 

* Cases like overlapped data from store and sidecar and 1:1 duplicates are now optimized by removal of exact duplicates as soon as it's possible, on proxy level (fanout receive). This case was highly visible on GitLab repro data and exists in most of the Thanos setups. cc @SuperQ 
* Also this adds logic for merging the same series together on proxy level instead of `select`. This allows for better dedup efficiency.

Signed-off-by: Bartlomiej Plotka <bwplotka@gmail.com>