@bwplotka 2020-05-13
As expected there is some overhead of this. 3x CPU 1.5x Mem allocs. Overhead was expected but maybe not that significant.

NOTE: Dataset does not have ANY duplicates, so we do in this function MORE, to do later, less (: 

```
name                                       old time/op    new time/op    delta
MergedSeriesSet/series=100,blocks=1-12        377µs ± 0%     428µs ± 0%   +13.52%
MergedSeriesSet/series=100,blocks=2-12        712µs ± 0%    3498µs ± 0%  +391.34%
MergedSeriesSet/series=100,blocks=4-12       1.46ms ± 0%    7.34ms ± 0%  +402.86%
MergedSeriesSet/series=100,blocks=8-12       3.08ms ± 0%   16.33ms ± 0%  +429.53%
MergedSeriesSet/series=100,blocks=16-12      6.31ms ± 0%   35.49ms ± 0%  +462.51%
MergedSeriesSet/series=100,blocks=32-12      13.3ms ± 0%    66.4ms ± 0%  +398.16%
MergedSeriesSet/series=1000,blocks=1-12      3.26ms ± 0%    3.40ms ± 0%    +4.33%
MergedSeriesSet/series=1000,blocks=2-12      8.12ms ± 0%   24.54ms ± 0%  +202.20%
MergedSeriesSet/series=1000,blocks=4-12      18.4ms ± 0%    68.1ms ± 0%  +270.21%
MergedSeriesSet/series=1000,blocks=8-12      37.7ms ± 0%   146.3ms ± 0%  +287.73%
MergedSeriesSet/series=1000,blocks=16-12     80.3ms ± 0%   301.1ms ± 0%  +274.88%
MergedSeriesSet/series=1000,blocks=32-12      155ms ± 0%     609ms ± 0%  +293.73%
MergedSeriesSet/series=10000,blocks=1-12     38.8ms ± 0%    34.5ms ± 0%   -11.14%
MergedSeriesSet/series=10000,blocks=2-12     87.6ms ± 0%   234.8ms ± 0%  +168.05%
MergedSeriesSet/series=10000,blocks=4-12      172ms ± 0%     607ms ± 0%  +253.76%
MergedSeriesSet/series=10000,blocks=8-12      419ms ± 0%    1377ms ± 0%  +228.45%
MergedSeriesSet/series=10000,blocks=16-12     821ms ± 0%    2849ms ± 0%  +247.14%
MergedSeriesSet/series=10000,blocks=32-12     1.83s ± 0%     5.89s ± 0%  +221.61%
MergedSeriesSet/series=20000,blocks=1-12      100ms ± 0%      69ms ± 0%   -31.37%
MergedSeriesSet/series=20000,blocks=2-12      199ms ± 0%     463ms ± 0%  +132.60%
MergedSeriesSet/series=20000,blocks=4-12      435ms ± 0%    1226ms ± 0%  +181.92%
MergedSeriesSet/series=20000,blocks=8-12      830ms ± 0%    2717ms ± 0%  +227.25%
MergedSeriesSet/series=20000,blocks=16-12     1.72s ± 0%     5.79s ± 0%  +236.17%
MergedSeriesSet/series=20000,blocks=32-12     3.41s ± 0%    11.81s ± 0%  +246.02%

name                                       old alloc/op   new alloc/op   delta
MergedSeriesSet/series=100,blocks=1-12        238kB ± 0%     238kB ± 0%    +0.00%
MergedSeriesSet/series=100,blocks=2-12        501kB ± 0%     708kB ± 0%   +41.25%
MergedSeriesSet/series=100,blocks=4-12       1.05MB ± 0%    1.65MB ± 0%   +56.43%
MergedSeriesSet/series=100,blocks=8-12       2.21MB ± 0%    3.53MB ± 0%   +59.71%
MergedSeriesSet/series=100,blocks=16-12      4.62MB ± 0%    7.29MB ± 0%   +57.69%
MergedSeriesSet/series=100,blocks=32-12      9.65MB ± 0%   14.80MB ± 0%   +53.35%
MergedSeriesSet/series=1000,blocks=1-12      2.35MB ± 0%    2.35MB ± 0%    +0.00%
MergedSeriesSet/series=1000,blocks=2-12      4.96MB ± 0%    7.02MB ± 0%   +41.72%
MergedSeriesSet/series=1000,blocks=4-12      10.4MB ± 0%    16.4MB ± 0%   +57.00%
MergedSeriesSet/series=1000,blocks=8-12      21.9MB ± 0%    35.0MB ± 0%   +60.24%
MergedSeriesSet/series=1000,blocks=16-12     45.8MB ± 0%    72.4MB ± 0%   +58.13%
MergedSeriesSet/series=1000,blocks=32-12     95.7MB ± 0%   147.1MB ± 0%   +53.78%
MergedSeriesSet/series=10000,blocks=1-12     25.0MB ± 0%    25.0MB ± 0%    +0.00%
MergedSeriesSet/series=10000,blocks=2-12     52.6MB ± 0%    73.3MB ± 0%   +39.21%
MergedSeriesSet/series=10000,blocks=4-12      110MB ± 0%     170MB ± 0%   +53.77%
MergedSeriesSet/series=10000,blocks=8-12      231MB ± 0%     363MB ± 0%   +56.99%
MergedSeriesSet/series=10000,blocks=16-12     483MB ± 0%     749MB ± 0%   +55.13%
MergedSeriesSet/series=10000,blocks=32-12    1.01GB ± 0%    1.52GB ± 0%   +51.12%
MergedSeriesSet/series=20000,blocks=1-12     50.2MB ± 0%    50.2MB ± 0%    -0.00%
MergedSeriesSet/series=20000,blocks=2-12      106MB ± 0%     147MB ± 0%   +39.11%
MergedSeriesSet/series=20000,blocks=4-12      221MB ± 0%     340MB ± 0%   +53.64%
MergedSeriesSet/series=20000,blocks=8-12      463MB ± 0%     727MB ± 0%   +56.86%
MergedSeriesSet/series=20000,blocks=16-12     967MB ± 0%    1499MB ± 0%   +55.02%
MergedSeriesSet/series=20000,blocks=32-12    2.02GB ± 0%    3.05GB ± 0%   +51.02%

name                                       old allocs/op  new allocs/op  delta
MergedSeriesSet/series=100,blocks=1-12        2.11k ± 0%     2.11k ± 0%     0.00%
MergedSeriesSet/series=100,blocks=2-12        4.31k ± 0%    11.41k ± 0%  +164.62%
MergedSeriesSet/series=100,blocks=4-12        8.72k ± 0%    30.02k ± 0%  +244.28%
MergedSeriesSet/series=100,blocks=8-12        17.5k ± 0%     67.2k ± 0%  +283.49%
MergedSeriesSet/series=100,blocks=16-12       35.2k ± 0%    141.7k ± 0%  +302.97%
MergedSeriesSet/series=100,blocks=32-12       70.4k ± 0%    290.5k ± 0%  +312.64%
MergedSeriesSet/series=1000,blocks=1-12       20.9k ± 0%     20.9k ± 0%     0.00%
MergedSeriesSet/series=1000,blocks=2-12       42.9k ± 0%    113.9k ± 0%  +165.56%
MergedSeriesSet/series=1000,blocks=4-12       86.8k ± 0%    299.8k ± 0%  +245.49%
MergedSeriesSet/series=1000,blocks=8-12        175k ± 0%      672k ± 0%  +284.76%
MergedSeriesSet/series=1000,blocks=16-12       350k ± 0%     1415k ± 0%  +304.23%
MergedSeriesSet/series=1000,blocks=32-12       701k ± 0%     2902k ± 0%  +313.93%
MergedSeriesSet/series=10000,blocks=1-12       209k ± 0%      209k ± 0%     0.00%
MergedSeriesSet/series=10000,blocks=2-12       428k ± 0%     1139k ± 0%  +165.70%
MergedSeriesSet/series=10000,blocks=4-12       867k ± 0%     2997k ± 0%  +245.69%
MergedSeriesSet/series=10000,blocks=8-12      1.74M ± 0%     6.71M ± 0%  +284.99%
MergedSeriesSet/series=10000,blocks=16-12     3.50M ± 0%    14.15M ± 0%  +304.47%
MergedSeriesSet/series=10000,blocks=32-12     7.01M ± 0%    29.02M ± 0%  +314.17%
MergedSeriesSet/series=20000,blocks=1-12       418k ± 0%      418k ± 0%     0.00%
MergedSeriesSet/series=20000,blocks=2-12       857k ± 0%     2277k ± 0%  +165.74%
MergedSeriesSet/series=20000,blocks=4-12      1.73M ± 0%     5.99M ± 0%  +245.74%
MergedSeriesSet/series=20000,blocks=8-12      3.49M ± 0%    13.43M ± 0%  +285.05%
MergedSeriesSet/series=20000,blocks=16-12     6.99M ± 0%    28.29M ± 0%  +304.54%
MergedSeriesSet/series=20000,blocks=32-12     14.0M ± 0%     58.0M ± 0%  +314.24%
```