@bwplotka 2020-05-13
Most of the mem comes from string comparison and marshalling... I guess we could do better :thinking: 

![image](https://user-images.githubusercontent.com/6950331/81863158-a96c4480-9562-11ea-9d3f-4816e441af19.png)