@brancz 2020-05-14
I think the strategy seems sane, but could we add a couple of benchmarks that prove the theory that this does indeed make the typical setup faster?