@bwplotka 2020-05-14
Ok, I know what's wrong (: 

![image](https://user-images.githubusercontent.com/6950331/81957493-8eeaa780-9604-11ea-9a06-25b8d6c2c5ff.png)

Looks like text comparison sucks.