@bwplotka 2020-05-16
Ok I optimized a lot and I think it's good to go. I can think of more optimizations, but it looks good.

```
benchstat -delta-test none _dev/mergesieres_set.out _dev/mergesieres_set3.out
name                                       old time/op    new time/op    delta
MergedSeriesSet/series=100,blocks=1-12        377µs ± 0%     398µs ± 0%    +5.61%
MergedSeriesSet/series=100,blocks=2-12        712µs ± 0%    1262µs ± 0%   +77.28%
MergedSeriesSet/series=100,blocks=4-12       1.46ms ± 0%    2.92ms ± 0%  +100.01%
MergedSeriesSet/series=100,blocks=8-12       3.08ms ± 0%    6.44ms ± 0%  +108.71%
MergedSeriesSet/series=100,blocks=16-12      6.31ms ± 0%   12.69ms ± 0%  +101.19%
MergedSeriesSet/series=100,blocks=32-12      13.3ms ± 0%    19.1ms ± 0%   +43.00%
MergedSeriesSet/series=1000,blocks=1-12      3.26ms ± 0%    4.53ms ± 0%   +39.15%
MergedSeriesSet/series=1000,blocks=2-12      8.12ms ± 0%   11.32ms ± 0%   +39.45%
MergedSeriesSet/series=1000,blocks=4-12      18.4ms ± 0%    19.5ms ± 0%    +5.94%
MergedSeriesSet/series=1000,blocks=8-12      37.7ms ± 0%    40.4ms ± 0%    +7.13%
MergedSeriesSet/series=1000,blocks=16-12     80.3ms ± 0%    83.4ms ± 0%    +3.87%
MergedSeriesSet/series=1000,blocks=32-12      155ms ± 0%     172ms ± 0%   +11.13%
MergedSeriesSet/series=10000,blocks=1-12     38.8ms ± 0%    27.2ms ± 0%   -29.93%
MergedSeriesSet/series=10000,blocks=2-12     87.6ms ± 0%    83.6ms ± 0%    -4.54%
MergedSeriesSet/series=10000,blocks=4-12      172ms ± 0%     202ms ± 0%   +17.69%
MergedSeriesSet/series=10000,blocks=8-12      419ms ± 0%     441ms ± 0%    +5.21%
MergedSeriesSet/series=10000,blocks=16-12     821ms ± 0%     891ms ± 0%    +8.54%
MergedSeriesSet/series=10000,blocks=32-12     1.83s ± 0%     1.80s ± 0%    -1.62%
MergedSeriesSet/series=20000,blocks=1-12      100ms ± 0%      56ms ± 0%   -44.11%
MergedSeriesSet/series=20000,blocks=2-12      199ms ± 0%     168ms ± 0%   -15.49%
MergedSeriesSet/series=20000,blocks=4-12      435ms ± 0%     395ms ± 0%    -9.11%
MergedSeriesSet/series=20000,blocks=8-12      830ms ± 0%     860ms ± 0%    +3.57%
MergedSeriesSet/series=20000,blocks=16-12     1.72s ± 0%     1.78s ± 0%    +3.29%
MergedSeriesSet/series=20000,blocks=32-12     3.41s ± 0%     3.61s ± 0%    +5.82%

name                                       old alloc/op   new alloc/op   delta
MergedSeriesSet/series=100,blocks=1-12        238kB ± 0%     238kB ± 0%     0.00%
MergedSeriesSet/series=100,blocks=2-12        501kB ± 0%     568kB ± 0%   +13.41%
MergedSeriesSet/series=100,blocks=4-12       1.05MB ± 0%    1.23MB ± 0%   +16.71%
MergedSeriesSet/series=100,blocks=8-12       2.21MB ± 0%    2.55MB ± 0%   +15.50%
MergedSeriesSet/series=100,blocks=16-12      4.62MB ± 0%    5.19MB ± 0%   +12.39%
MergedSeriesSet/series=100,blocks=32-12      9.65MB ± 0%   10.48MB ± 0%    +8.59%
MergedSeriesSet/series=1000,blocks=1-12      2.35MB ± 0%    2.35MB ± 0%     0.00%
MergedSeriesSet/series=1000,blocks=2-12      4.96MB ± 0%    5.63MB ± 0%   +13.56%
MergedSeriesSet/series=1000,blocks=4-12      10.4MB ± 0%    12.2MB ± 0%   +16.89%
MergedSeriesSet/series=1000,blocks=8-12      21.9MB ± 0%    25.3MB ± 0%   +15.66%
MergedSeriesSet/series=1000,blocks=16-12     45.8MB ± 0%    51.5MB ± 0%   +12.51%
MergedSeriesSet/series=1000,blocks=32-12     95.7MB ± 0%   104.0MB ± 0%    +8.66%
MergedSeriesSet/series=10000,blocks=1-12     25.0MB ± 0%    25.0MB ± 0%    -0.00%
MergedSeriesSet/series=10000,blocks=2-12     52.6MB ± 0%    59.4MB ± 0%   +12.76%
MergedSeriesSet/series=10000,blocks=4-12      110MB ± 0%     128MB ± 0%   +15.94%
MergedSeriesSet/series=10000,blocks=8-12      231MB ± 0%     265MB ± 0%   +14.82%
MergedSeriesSet/series=10000,blocks=16-12     483MB ± 0%     540MB ± 0%   +11.87%
MergedSeriesSet/series=10000,blocks=32-12    1.01GB ± 0%    1.09GB ± 0%    +8.24%
MergedSeriesSet/series=20000,blocks=1-12     50.2MB ± 0%    50.2MB ± 0%    -0.00%
MergedSeriesSet/series=20000,blocks=2-12      106MB ± 0%     119MB ± 0%   +12.73%
MergedSeriesSet/series=20000,blocks=4-12      221MB ± 0%     257MB ± 0%   +15.90%
MergedSeriesSet/series=20000,blocks=8-12      463MB ± 0%     532MB ± 0%   +14.79%
MergedSeriesSet/series=20000,blocks=16-12     967MB ± 0%    1082MB ± 0%   +11.84%
MergedSeriesSet/series=20000,blocks=32-12    2.02GB ± 0%    2.18GB ± 0%    +8.22%

name                                       old allocs/op  new allocs/op  delta
MergedSeriesSet/series=100,blocks=1-12        2.11k ± 0%     2.11k ± 0%     0.00%
MergedSeriesSet/series=100,blocks=2-12        4.31k ± 0%     7.31k ± 0%   +69.56%
MergedSeriesSet/series=100,blocks=4-12        8.72k ± 0%    17.72k ± 0%  +103.21%
MergedSeriesSet/series=100,blocks=8-12        17.5k ± 0%     38.5k ± 0%  +119.77%
MergedSeriesSet/series=100,blocks=16-12       35.2k ± 0%     80.2k ± 0%  +127.99%
MergedSeriesSet/series=100,blocks=32-12       70.4k ± 0%    163.4k ± 0%  +132.09%
MergedSeriesSet/series=1000,blocks=1-12       20.9k ± 0%     20.9k ± 0%     0.00%
MergedSeriesSet/series=1000,blocks=2-12       42.9k ± 0%     72.9k ± 0%   +69.95%
MergedSeriesSet/series=1000,blocks=4-12       86.8k ± 0%    176.8k ± 0%  +103.72%
MergedSeriesSet/series=1000,blocks=8-12        175k ± 0%      385k ± 0%  +120.32%
MergedSeriesSet/series=1000,blocks=16-12       350k ± 0%      800k ± 0%  +128.55%
MergedSeriesSet/series=1000,blocks=32-12       701k ± 0%     1631k ± 0%  +132.64%
MergedSeriesSet/series=10000,blocks=1-12       209k ± 0%      209k ± 0%     0.00%
MergedSeriesSet/series=10000,blocks=2-12       428k ± 0%      728k ± 0%   +70.01%
MergedSeriesSet/series=10000,blocks=4-12       867k ± 0%     1767k ± 0%  +103.81%
MergedSeriesSet/series=10000,blocks=8-12      1.74M ± 0%     3.84M ± 0%  +120.42%
MergedSeriesSet/series=10000,blocks=16-12     3.50M ± 0%     8.00M ± 0%  +128.65%
MergedSeriesSet/series=10000,blocks=32-12     7.01M ± 0%    16.31M ± 0%  +132.75%
MergedSeriesSet/series=20000,blocks=1-12       418k ± 0%      418k ± 0%     0.00%
MergedSeriesSet/series=20000,blocks=2-12       857k ± 0%     1457k ± 0%   +70.03%
MergedSeriesSet/series=20000,blocks=4-12      1.73M ± 0%     3.53M ± 0%  +103.83%
MergedSeriesSet/series=20000,blocks=8-12      3.49M ± 0%     7.69M ± 0%  +120.44%
MergedSeriesSet/series=20000,blocks=16-12     6.99M ± 0%    15.99M ± 0%  +128.68%
MergedSeriesSet/series=20000,blocks=32-12     14.0M ± 0%     32.6M ± 0%  +132.78%
```

NOTE: I just noticed this is the *full* overlapping benchmark. So it means 100% chunks in those "blocks" are overlapping. That's why comparison was so crucial (: This is normally a rare operation though. I extended tests with non overlap as well.