@bwplotka 2020-05-16
Ok last version uses LESS resources and is FASTER and adds MORE features =D

```
benchstat -delta-test none _dev/mergesieres_set.out _dev/mergesieres_set4.out
name                                       old time/op    new time/op    delta
MergedSeriesSet/series=100,blocks=1-12        377µs ± 0%     562µs ± 0%  +48.92%
MergedSeriesSet/series=100,blocks=2-12        712µs ± 0%     971µs ± 0%  +36.46%
MergedSeriesSet/series=100,blocks=4-12       1.46ms ± 0%    1.80ms ± 0%  +23.26%
MergedSeriesSet/series=100,blocks=8-12       3.08ms ± 0%    3.10ms ± 0%   +0.58%
MergedSeriesSet/series=100,blocks=16-12      6.31ms ± 0%    5.89ms ± 0%   -6.67%
MergedSeriesSet/series=100,blocks=32-12      13.3ms ± 0%    10.3ms ± 0%  -22.38%
MergedSeriesSet/series=1000,blocks=1-12      3.26ms ± 0%    2.82ms ± 0%  -13.31%
MergedSeriesSet/series=1000,blocks=2-12      8.12ms ± 0%    5.87ms ± 0%  -27.74%
MergedSeriesSet/series=1000,blocks=4-12      18.4ms ± 0%    11.8ms ± 0%  -35.74%
MergedSeriesSet/series=1000,blocks=8-12      37.7ms ± 0%    24.0ms ± 0%  -36.40%
MergedSeriesSet/series=1000,blocks=16-12     80.3ms ± 0%    49.8ms ± 0%  -37.96%
MergedSeriesSet/series=1000,blocks=32-12      155ms ± 0%     100ms ± 0%  -35.04%
MergedSeriesSet/series=10000,blocks=1-12     38.8ms ± 0%    26.6ms ± 0%  -31.30%
MergedSeriesSet/series=10000,blocks=2-12     87.6ms ± 0%    58.5ms ± 0%  -33.26%
MergedSeriesSet/series=10000,blocks=4-12      172ms ± 0%     124ms ± 0%  -27.96%
MergedSeriesSet/series=10000,blocks=8-12      419ms ± 0%     258ms ± 0%  -38.54%
MergedSeriesSet/series=10000,blocks=16-12     821ms ± 0%     536ms ± 0%  -34.69%
MergedSeriesSet/series=10000,blocks=32-12     1.83s ± 0%     1.09s ± 0%  -40.52%
MergedSeriesSet/series=20000,blocks=1-12      100ms ± 0%      55ms ± 0%  -45.65%
MergedSeriesSet/series=20000,blocks=2-12      199ms ± 0%     120ms ± 0%  -39.70%
MergedSeriesSet/series=20000,blocks=4-12      435ms ± 0%     255ms ± 0%  -41.33%
MergedSeriesSet/series=20000,blocks=8-12      830ms ± 0%     527ms ± 0%  -36.48%
MergedSeriesSet/series=20000,blocks=16-12     1.72s ± 0%     1.09s ± 0%  -36.85%
MergedSeriesSet/series=20000,blocks=32-12     3.41s ± 0%     2.20s ± 0%  -35.66%

name                                       old alloc/op   new alloc/op   delta
MergedSeriesSet/series=100,blocks=1-12        238kB ± 0%     238kB ± 0%   +0.00%
MergedSeriesSet/series=100,blocks=2-12        501kB ± 0%     501kB ± 0%   +0.00%
MergedSeriesSet/series=100,blocks=4-12       1.05MB ± 0%    1.03MB ± 0%   -2.43%
MergedSeriesSet/series=100,blocks=8-12       2.21MB ± 0%    2.08MB ± 0%   -5.80%
MergedSeriesSet/series=100,blocks=16-12      4.62MB ± 0%    4.19MB ± 0%   -9.42%
MergedSeriesSet/series=100,blocks=32-12      9.65MB ± 0%    8.40MB ± 0%  -13.00%
MergedSeriesSet/series=1000,blocks=1-12      2.35MB ± 0%    2.35MB ± 0%   -0.00%
MergedSeriesSet/series=1000,blocks=2-12      4.96MB ± 0%    4.96MB ± 0%   -0.00%
MergedSeriesSet/series=1000,blocks=4-12      10.4MB ± 0%    10.2MB ± 0%   -2.46%
MergedSeriesSet/series=1000,blocks=8-12      21.9MB ± 0%    20.6MB ± 0%   -5.85%
MergedSeriesSet/series=1000,blocks=16-12     45.8MB ± 0%    41.4MB ± 0%   -9.50%
MergedSeriesSet/series=1000,blocks=32-12     95.7MB ± 0%    83.1MB ± 0%  -13.11%
MergedSeriesSet/series=10000,blocks=1-12     25.0MB ± 0%    25.0MB ± 0%    0.00%
MergedSeriesSet/series=10000,blocks=2-12     52.6MB ± 0%    52.6MB ± 0%   +0.00%
MergedSeriesSet/series=10000,blocks=4-12      110MB ± 0%     108MB ± 0%   -2.32%
MergedSeriesSet/series=10000,blocks=8-12      231MB ± 0%     218MB ± 0%   -5.54%
MergedSeriesSet/series=10000,blocks=16-12     483MB ± 0%     439MB ± 0%   -9.02%
MergedSeriesSet/series=10000,blocks=32-12    1.01GB ± 0%    0.88GB ± 0%  -12.47%
MergedSeriesSet/series=20000,blocks=1-12     50.2MB ± 0%    50.2MB ± 0%   -0.00%
MergedSeriesSet/series=20000,blocks=2-12      106MB ± 0%     106MB ± 0%   +0.00%
MergedSeriesSet/series=20000,blocks=4-12      221MB ± 0%     216MB ± 0%   -2.31%
MergedSeriesSet/series=20000,blocks=8-12      463MB ± 0%     438MB ± 0%   -5.53%
MergedSeriesSet/series=20000,blocks=16-12     967MB ± 0%     880MB ± 0%   -9.00%
MergedSeriesSet/series=20000,blocks=32-12    2.02GB ± 0%    1.77GB ± 0%  -12.44%

name                                       old allocs/op  new allocs/op  delta
MergedSeriesSet/series=100,blocks=1-12        2.11k ± 0%     2.11k ± 0%    0.00%
MergedSeriesSet/series=100,blocks=2-12        4.31k ± 0%     4.31k ± 0%    0.00%
MergedSeriesSet/series=100,blocks=4-12        8.72k ± 0%     8.72k ± 0%    0.00%
MergedSeriesSet/series=100,blocks=8-12        17.5k ± 0%     17.5k ± 0%    0.00%
MergedSeriesSet/series=100,blocks=16-12       35.2k ± 0%     35.2k ± 0%    0.00%
MergedSeriesSet/series=100,blocks=32-12       70.4k ± 0%     70.4k ± 0%    0.00%
MergedSeriesSet/series=1000,blocks=1-12       20.9k ± 0%     20.9k ± 0%    0.00%
MergedSeriesSet/series=1000,blocks=2-12       42.9k ± 0%     42.9k ± 0%    0.00%
MergedSeriesSet/series=1000,blocks=4-12       86.8k ± 0%     86.8k ± 0%    0.00%
MergedSeriesSet/series=1000,blocks=8-12        175k ± 0%      175k ± 0%    0.00%
MergedSeriesSet/series=1000,blocks=16-12       350k ± 0%      350k ± 0%    0.00%
MergedSeriesSet/series=1000,blocks=32-12       701k ± 0%      701k ± 0%    0.00%
MergedSeriesSet/series=10000,blocks=1-12       209k ± 0%      209k ± 0%    0.00%
MergedSeriesSet/series=10000,blocks=2-12       428k ± 0%      428k ± 0%    0.00%
MergedSeriesSet/series=10000,blocks=4-12       867k ± 0%      867k ± 0%    0.00%
MergedSeriesSet/series=10000,blocks=8-12      1.74M ± 0%     1.74M ± 0%    0.00%
MergedSeriesSet/series=10000,blocks=16-12     3.50M ± 0%     3.50M ± 0%    0.00%
MergedSeriesSet/series=10000,blocks=32-12     7.01M ± 0%     7.01M ± 0%    0.00%
MergedSeriesSet/series=20000,blocks=1-12       418k ± 0%      418k ± 0%    0.00%
MergedSeriesSet/series=20000,blocks=2-12       857k ± 0%      857k ± 0%    0.00%
MergedSeriesSet/series=20000,blocks=4-12      1.73M ± 0%     1.73M ± 0%    0.00%
MergedSeriesSet/series=20000,blocks=8-12      3.49M ± 0%     3.49M ± 0%    0.00%
MergedSeriesSet/series=20000,blocks=16-12     6.99M ± 0%     6.99M ± 0%    0.00%
MergedSeriesSet/series=20000,blocks=32-12     14.0M ± 0%     14.0M ± 0%    0.00%
```
Considering this done!